requirements.txt
Flask
Flask-Cors
mysql-connector-python
python-dotenv
Werkzeug
gunicorn 
requests
flask_session
pandas
scikit-learn
google-generativeai
googlemaps
ortools